management-paper-agent/requirements.txt
openai>=1.40.0
PyYAML>=6.0.1
requests>=2.32.0
